Renamed try-index.js -> index.js
