+++ lib/ldp-middleware.js
@@ -9,7 +9,7 @@ var post = require('./handlers/post')
 var put = require('./handlers/put')
 var del = require('./handlers/delete')
 var patch = require('./handlers/patch')
-var index = require('./handlers/try-index')
+var index = require('./handlers/index')
 var errorPages = require('./handlers/error-pages')
 
 function LdpMiddleware (corsSettings) {
